Rafactorization, renaming current game instance to ClientLocalGame
