+++ pong_standard/src/client/DistantBall.ts
@@ -1,0 +1,5 @@
+import { Ball } from "./Ball";
+
+export class DistantBall {
+
+}
