+++ pong_standard/src/client/DistantBar.ts
@@ -1,0 +1,6 @@
+import * as THREE from 'three'
+import { Bar } from './Bar';
+
+export class DistantBar extends Bar {
+    update(elapsedTime: number) {}
+}
